Problem_Submission_Service/requirements.txt
asgiref==3.8.1
Django==4.2.11
sqlparse==0.5.0
typing-extensions==4.11.0
djangorestframework
psycopg2-binary
ortools>=9.0
requests